post replacement: move presenter methods to post_replacement.rb.
